requirements.txt
amqp==1.4.6
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
blessings==1.6
bpython==0.14.1
celery==3.1.17
certifi==14.5.14
curtsies==0.1.18
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.5
django-allauth==0.19.1
django-audiofield==0.6.4
django-bootstrap3==5.1.1
django-celery==3.1.16
django-debug-toolbar==1.3.0
django-extensions==1.5.1
django-jquery==1.9.1
django-mailgun==0.2.2
django-phonenumber-field==0.7.2
django-toolbelt==0.0.1
greenlet==0.4.5
gunicorn==19.3.0
httplib2==0.9
ipdb==0.8
ipython==3.0.0
kombu==3.0.24
oauthlib==0.7.2
phonenumbers==7.0.2
psycopg2==2.6
pudb==2015.2
pycrypto==2.6.1
Pygments==2.0.2
python-dateutil==2.4.1
python-openid==2.2.5
pytz==2014.10
redis==2.10.3
requests==2.5.3
requests-oauthlib==0.4.2
six==1.9.0
sockjs-tornado==1.0.1
sqlparse==0.1.14
static3==0.5.1
SwampDragon==0.4.1.2
tornado==4.1
tornado-redis==2.4.18
twilio==3.7.3
urwid==1.3.0
Werkzeug==0.10.1
whitenoise==1.0.6
